+++ app.js
@@ -1,15 +1,13 @@
 const express = require("express");
 const cors = require("cors");
-const http = require("http");
-const { router: postsRoutes } = require("./routes/routes");
-const sslRedirect = require('express-sslify');
-
-const PORT = process.env.PORT || 5000;
+const sslRedirect = require("express-sslify");
+const { postsRoutes } = require("./routes");
+const errorHandler = require("./middlewares/errorHandler");
 
 const app = express();
-const server = http.createServer(app);
 
-
+// Redirect HTTP to HTTPS
+app.use(sslRedirect.HTTPS({ trustProtoHeader: true }));
 
 // Middleware setup
 app.use(express.json());
@@ -17,18 +15,7 @@ app.use(cors());
 app.use("/assets", express.static("assets"));
 app.use("/posts", postsRoutes);
 
-// Redirect HTTP to HTTPS
-app.use(sslRedirect.HTTPS({ trustProtoHeader: true }));
+// Centralized Error Handling Middleware
+app.use(errorHandler);
 
-// Error Handling
-process.on('uncaughtException', (err) => {
-  console.error('Uncaught Exception:', err);
-});
-
-process.on('unhandledRejection', (reason, promise) => {
-  console.error('Unhandled Rejection at:', promise, 'reason:', reason);
-});
-
-server.listen(PORT, () => {
-  console.log(`Server is running on port ${PORT}`);
-});
+module.exports = app;
