+++ routes/index,js
@@ -1,0 +1,5 @@
+const postsRoutes = require("./posts");
+
+module.exports = {
+  postsRoutes,
+};
